fixed small data race. updated tests.
